+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.9.1
 
+* `CacheDict`: `try_delete` -> `invalidate`
+